Improve handling of "media_dirs" file extension config
This moves almost all of the processing for the "media_dirs" file setting
into the config system, with the aim of making it more robust while also
resolving some bugs. The only thing left for the file library backend to
do manually now is verify that the input directories actually exist as
folders.

The main bug this solves is an issue of whitespace handling around
paths, as detailed in #1965.

To make this possible, the Path config type has been altered to return
pathlib objects rather than strings.